Several bug and warning fixes
- Fixed improper exception throw syntax in RfgMaterial.h
- Fixed Packfile3::ExtractSingleFile() not having an explicit return statement in one path.
- Fixed memory leak in Peg10::Cleanup() due to logic error
- Fixed warnings in Vec2.h and Vec3.h about double to float conversions

+++ RfgTools++/RfgTools++/types/Vec2.h
@@ -40,7 +40,7 @@ struct Vec2
 
     f32 Distance(const Vec2& B) const
     {
-        return sqrt(pow(B.x - x, 2) + pow(B.y - y, 2));
+        return sqrtf(powf(B.x - x, 2.0f) + powf(B.y - y, 2.0f));
     }
 
     f32 Magnitude()
